remove ListFugue vs TreeFugue
