+++ crdt-benchmarks/benchmarks/fugue-simple/benchmark.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>Testing Tree-Fugue</title>
+    <title>Testing Fugue Simple</title>
   </head>
   <body>
     <script type="module" src="./dist/benchmark-browser.js"></script>
