+++ crdt-benchmarks/benchmarks/fugue-simple/package.json
@@ -1,20 +1,20 @@
 {
-  "name": "list-fugue-simple-bench",
+  "name": "fugue-simple-bench",
   "version": "1.0.0",
-  "description": "List-Fugue Simple benchmarks",
+  "description": "Fugue Simple benchmarks",
   "main": "./run.js",
   "type": "module",
   "scripts": {
     "clean": "rm -rf dist",
     "start": "([ -f \"./dist/bundle.js\" ] || npm run dist) && node --expose-gc run.js && npm run table",
-    "table": "echo 'List-Fugue Simple results: \n\n' && node ../../bin/render-table.js ../results.json 6000 list-fugue-simple",
+    "table": "echo 'Fugue Simple results: \n\n' && node ../../bin/render-table.js ../results.json 6000 fugue-simple",
     "dist": "npm run clean && rollup -c && gzip --keep dist/*"
   },
-  "author": "Matthew Weidner, Seph Gentle, and Martin Kleppmann",
+  "author": "Matthew Weidner and Martin Kleppmann",
   "license": "MIT",
   "dependencies": {
     "@collabs/collabs": "0.6.1",
     "seedrandom": "^3.0.5",
-    "list-fugue-simple": "1.0.0"
+    "fugue-simple": "1.0.0"
   }
 }
