+++ crdt-benchmarks/benchmarks/fugue-simple/run.js
