+++ crdt-benchmarks/benchmarks/fugue/benchmark.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>Testing Tree-Fugue Simple</title>
+    <title>Testing Fugue</title>
   </head>
   <body>
     <script type="module" src="./dist/benchmark-browser.js"></script>
