+++ crdt-benchmarks/benchmarks/fugue/package.json
@@ -1,21 +1,21 @@
 {
-  "name": "tree-fugue-bench",
+  "name": "fugue-bench",
   "version": "1.0.0",
-  "description": "Tree-Fugue benchmarks",
+  "description": "Fugue benchmarks",
   "main": "./run.js",
   "type": "module",
   "scripts": {
     "clean": "rm -rf dist",
     "start": "([ -f \"./dist/bundle.js\" ] || npm run dist) && node --expose-gc run.js && npm run table",
-    "table": "echo 'Tree-Fugue results: \n\n' && node ../../bin/render-table.js ../results.json 6000 tree-fugue",
+    "table": "echo 'Fugue results: \n\n' && node ../../bin/render-table.js ../results.json 6000 fugue",
     "B4x100": "([ -f \"./dist/bundle.js\" ] || npm run dist) && node --expose-gc run.js B4x100 && npm run table",
     "dist": "npm run clean && rollup -c && gzip --keep dist/*"
   },
-  "author": "Matthew Weidner, Seph Gentle, and Martin Kleppmann",
+  "author": "Matthew Weidner and Martin Kleppmann",
   "license": "MIT",
   "dependencies": {
     "@collabs/collabs": "0.6.1",
     "seedrandom": "^3.0.5",
-    "tree-fugue": "1.0.0"
+    "fugue": "1.0.0"
   }
 }
