--- crdt-benchmarks/benchmarks/list-fugue-simple/benchmark.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Testing List-Fugue Simple</title>
-  </head>
-  <body>
-    <script type="module" src="./dist/benchmark-browser.js"></script>
-  </body>
-</html>
