--- crdt-benchmarks/benchmarks/tree-fugue/rollup.config.js
@@ -1,3 +1,0 @@
-import rollupConfig from '../../js-lib/rollup-helper.js'
-
-export default rollupConfig
