+++ fugue-simple/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "tree-fugue-simple",
+  "name": "fugue-simple",
   "version": "1.0.0",
-  "author": "Matthew Weidner, Seph Gentle, and Martin Kleppmann",
+  "author": "Matthew Weidner and Martin Kleppmann",
   "main": "dist/index.js",
   "type": "module",
   "types": "dist/index.d.ts",
