+++ fugue/src/list_position_source/array_list_item_manager.ts
