+++ fugue/src/list_position_source/number_list_item_manager.ts
