+++ fugue/src/list_position_source/serialization.ts
