+++ fugue/src/list_position_source/string_list_item_manager.ts
