+++ fugue/src/tree_fugue_array.proto
