--- list-fugue-simple/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "list-fugue-simple",
-  "version": "1.0.0",
-  "author": "Matthew Weidner, Seph Gentle, and Martin Kleppmann",
-  "main": "dist/index.js",
-  "type": "module",
-  "types": "dist/index.d.ts",
-  "dependencies": {
-    "@collabs/collabs": "0.6.1",
-    "pako": "^2.1.0",
-    "tslib": "^2.4.1"
-  },
-  "devDependencies": {
-    "@types/pako": "^2.0.0",
-    "typescript": "^4.9.4"
-  },
-  "scripts": {
-    "build": "tsc",
-    "clean": "rm -rf dist"
-  }
-}
