--- tree-fugue-simple/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "strict": true,
-    "target": "ES2020",
-    "module": "ES2020",
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    "sourceMap": true,
-    "importHelpers": true,
-    "declaration": true,
-    "outDir": "dist"
-  },
-  "include": ["src"]
-}
