--- yjs-interleave/package.json
@@ -1,12 +1,0 @@
-{
-  "name": "yjs-interleave",
-  "description": "Yjs interleaving example from the paper",
-  "version": "1.0.0",
-  "dependencies": {
-    "yjs": "13.5.44"
-  },
-  "scripts": {
-    "start": "node index.js",
-    "clean": ""
-  }
-}
